Fixed bug, much more stable now. Aesthetic changes to animation, cleaned up PS3D, added new 3D geometry and dynamics files.
Before in a single time stop, the forces acting a single particle were summed, that particles new state was calculated (overwriting the old state), then the forces and state of the next particle were summed, etc. This led to energy instability and a lack of symetry.

Now, in a single time step, the forces acting on ALL the particles at the begining of the time step are calculated, and the new positions of ALL the particles are then calculated. Symetry and energy conservation are much more consistent now.